Minor bugfix:
The drainInbox method could potentially lose messages if there was an 
error adding a message to the store, because it was being removed from the
maildir before being added to the store.

I've moved the removal of the message from the maildir until after adding to 
the storage, to prevent this from happening.  I discovered this with a test
that tried to drain the same message twice, so the unittest uses this 
strategy.